Added a resource not found exception every where necessary

+++ ProductManagement/src/main/java/org/example/productmanagement/Services/CategoryService.java
@@ -2,6 +2,7 @@ package org.example.productmanagement.Services;
 
 import org.example.productmanagement.DTOs.CategoryDTO;
 import org.example.productmanagement.Exceptions.ResourceAlreadyExists;
+import org.example.productmanagement.Exceptions.ResourceNotFound;
 import org.example.productmanagement.Models.Category;
 import org.example.productmanagement.Repositories.CategoryRepository;
 import org.springframework.stereotype.Service;
@@ -19,6 +20,7 @@ public class CategoryService {
     }
 
     public CategoryDTO createCategory(CategoryDTO categoryDTO) {
+
         //Check if a product already exists
         if (categoryRepo.existsByName(categoryDTO.getName())) {
             throw new ResourceAlreadyExists("Category with name '" + categoryDTO.getName() + "' already exists");
@@ -40,12 +42,18 @@ public class CategoryService {
                 .collect(Collectors.toList());
     }
 
-    public Optional<CategoryDTO> getCategoryById(int id) {
+    public CategoryDTO getCategoryById(int id) {
         return categoryRepo.findById(id)
-                .map (category -> new CategoryDTO(category.getName(),category.getDescription()));
+                .map (category -> new CategoryDTO(category.getName(),category.getDescription()))
+                .orElseThrow(() -> new ResourceNotFound("Category with id '" + id + "' was not found"));
     }
 
-    public Optional<CategoryDTO> updateCategory(int id, CategoryDTO categoryDTO) {
+    public CategoryDTO updateCategory(int id, CategoryDTO categoryDTO) {
+
+        //Check if a category id already exists
+        if (!categoryRepo.existsById(id)) {
+            throw new ResourceNotFound("Category with name '" + categoryDTO.getName() + "'  and id '"+ id + "' was not found");
+        }
 
         categoryRepo.findById(id)
                 .ifPresent(category -> {
@@ -58,6 +66,10 @@ public class CategoryService {
     }
 
     public void deleteCategory(int id) {
+        //Check if a category id already exists
+        if (!categoryRepo.existsById(id)) {
+            throw new ResourceNotFound("Category with id '"+ id + "' was not found");
+        }
         categoryRepo.deleteById(id);
     }
 }
